% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_a_pulgadas.R
\name{cm_a_pulgadas}
\alias{cm_a_pulgadas}
\title{Conversión de centímetros a pulgadas}
\usage{
cm_a_pulgadas(cm)
}
\arguments{
\item{cm}{Un valor numérico (o vector numérico) que representa la longitud en centímetros.}
}
\value{
Un valor numérico (o vector numérico) que representa la longitud en pulgadas.
}
\description{
Esta función convierte una medida de longitud expresada en centímetros a su equivalente en pulgadas.
}
\examples{
cm_a_pulgadas(10)
cm_a_pulgadas(c(10, 25.4, 100))
}
