% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_pulgada.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Conversion de pulgadas a centimetros}
\usage{
pulgadas_a_centimetros(medida_pulgada)
}
\arguments{
\item{medida_pulgada}{un valor en pulgadas.}
}
\value{
un valor o vector numérico con la conversion de las pulgadas
}
\description{
convierte las medidas de pulgadas a centimetros
}
\examples{
pulgadas_a_centimetros(30)

}
