% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(x = 1, y = 4)
}
\arguments{
\item{x}{un valor o vector numérico}

\item{y}{un valor o vector numérico}
}
\value{
Un valor o vector numérico que es la suma de \code{x} e \code{y}.
}
\description{
Suma dos numeros que deben ser positivos.
}
\examples{
suma(10,3)
suma(5)
}
