% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_a_pulgadas.R
\name{cm_a_pulgadas}
\alias{cm_a_pulgadas}
\title{Conversión de centímetros a pulgadas}
\usage{
cm_a_pulgadas(medida_cm)
}
\arguments{
\item{medida_cm}{Un número o vector numérico que representa la medida
convertida a pulgadas}
}
\value{
Un número o vector numérico con la medida convertida a pulgadas.
}
\description{
Convierte una medida dada en centímetros a su equivalente en pulgadas.
}
\examples{
cm_a_pulgadas(2.54) #Devuelve 1
cm_a_pulgadas(c(10,25)) # Convierte varios valores

}
