% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_a_pulgadas.R
\name{cm_a_pulgadas}
\alias{cm_a_pulgadas}
\title{Convierte centímetros a pulgadas}
\usage{
cm_a_pulgadas(cm)
}
\arguments{
\item{cm}{Un valor o vector numérico que representa longitudes en centímetros.}
}
\value{
Un valor numérico con la longitud equivalente en pulgadas.
}
\description{
Convierte un valor o vector numérico de longitudes expresadas en centímetros a su equivalente en pulgadas.
}
\examples{
cm_a_pulgadas(10)
cm_a_pulgadas(c(5, 15, 25))

}
