% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{datos}
\alias{datos}
\title{datos aleatorios}
\format{
vector de largo 1000.

\describe{
\item{datos}{númerico, de largo 1000.}
}
}
\usage{
datos
}
\description{
set de datos de ejemplo com datos aleatorios
}
\keyword{datasets}
