% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{datos}
\alias{datos}
\title{Datos aleatorios de ejemplo}
\format{
Un vector numérico de largo 1000.
}
\usage{
data(datos)
}
\description{
Set de datos de ejemplo con valores aleatorios.
}
\examples{
length(datos)
}
\keyword{datasets}
