% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{datos}
\alias{datos}
\title{Datos aleatorios}
\format{
vector de largo 1000.

\describe{
\item{datos}{numérico, de largo 1000.}
}
}
\usage{
datos
}
\description{
Set de datos de ejemplo con datos aleatorios
}
\keyword{datasets}
