% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitud.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Pulgadas a centímetros}
\usage{
pulgadas_a_centimetros(medida_pulgadas)
}
\arguments{
\item{medida_pulgadas}{un valor o vector numérico}
}
\value{
devuelve un valor numérico con la conversión de pulgadas a centímetros
}
\description{
Esta función convierte el número ingresado en pulgadas a centímetros
}
\examples{
pulgadas_a_centimetros(3)
pulgadas_a_centimetros(8)

}
