% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pulgadas.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Pulgadas a centimetros}
\usage{
pulgadas_a_centimetros(medida_pulgadas)
}
\arguments{
\item{medida_pulgadas}{Número o vector numérico medido en pulgadas}
}
\value{
Devuelve un numero o vector numérico medido en centímetros.
}
\description{
A partir de una medida en pulgadas la funcion la pasa a una medida en centímetros.
}
\examples{
pulgadas_a_centimetros(56)
pulgadas_a_centimetros(25.3)
}
