% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(a, b)
}
\arguments{
\item{a}{un valor o vector numérico.}

\item{b}{un valor o vector numérico.}
}
\value{
un valor o vector numérico con la suma de x más y.
}
\description{
suma dos numeros que deben ser positivo.
}
\examples{
suma(10,2)
suma(3,1)

}
