% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(x = 0, y = 0)
}
\arguments{
\item{x}{Un número positivo (no negativo).}

\item{y}{Un número positivo (no negativo).}
}
\value{
Un vector numérico.
}
\description{
Suma dos números numéricos y positivos.
}
\examples{
suma(1,1)
suma(3,8)
}
