% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos números}
\usage{
suma(x, y)
}
\arguments{
\item{x}{Un valor o vector numérico}

\item{y}{Un valor o vector numérico}
}
\value{
Un valor o vector numérico con la suma de x + y
}
\description{
Suma dos numeros que deben ser positivos
}
\examples{
suma(10,2)
suma (3, 1)
}
