% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un valor o vector numerico}

\item{y}{un valor o vector numerico}
}
\value{
un valor numerico con la suma de x e y
}
\description{
Suma dos numeros que deben ser positivos
}
\examples{
suma(10,2)
suma(3,1)

}
