% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos Números}
\usage{
suma(x = 10, y = 20)
}
\arguments{
\item{x}{Un valor o vector numérico.}

\item{y}{Un valor o vector numérico.}
}
\value{
Un valor o vector numérico con la suma de x más y.
}
\description{
Suma dos números que deben ser posistivos.
}
\examples{
suma(10,2)
suma(3,1)

}
