% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un valor o vector numérico.}

\item{y}{un valor o vector numérico.}
}
\value{
Un valor o vector numérico con la suma de x más y.
}
\description{
Suma dos números que deben ser positivos.
}
\examples{
suma(10, 2)
suma(1, 1)

}
